refactor(acor): extract Redis client creation to client.go
